+++ app/(user)/LikeDocsList.tsx
@@ -1,5 +1,5 @@
 import Accordion from "@/components/Accordion";
-import { ContributeDocsType } from "@/types/contributeDocs.interface";
+import { ContributeDocsType } from "@/types";
 import Link from "next/link";
 import { CLASSIFY } from "@/record";
 import * as styles from "./style.css";
