+++ components/Popular/style.css.ts
@@ -1,6 +1,6 @@
 import { style, styleVariants } from "@vanilla-extract/css";
 import { flex, font, theme } from "@/styles";
-import StyleVariantsType from "@/types/styleVariants.type";
+import { StyleVariantsType } from "@/types";
 import { popularAnimation } from "./keyframes.css";
 
 const containerBase = style({
